Created rankhospital.R function.
It takes 3 arguments: the 2-character abbreviated name of a state, an outcome ("heart attack", "heart failure", or "pneumonia"), and the ranking of a hospital in that state for that outcome.

It returns a character vector with the name of the hospital that has the ranking specified in the specified state and for the given outcome. Ties are solved by returning the first hospital name when sorted in alphabetical order.