Created rankall.R function.
It takes two arguments: an outcome ("heart attack", "heart attack", or "pneumonia") and a hospital ranking.

The function returns a 2-column data frame containing the hospital in each state that has the specified ranking. If a given hospital has no data for the specified outcome, it should be excluded. Ties are resolved by returning the first hospital when sorted in alphabetical order.